RHEL/Fedora Equivalent Dependencies
For anyone who would like to install the RHEL/Fedora equivalent dependencies.

NOTE: If one runs into the problem where running "./build.sh" shows Qemu's ./configure saying
"ERROR: "cc" either does not exist or does not work"
it's because of permission conflicts under ~/.ccache/4/9 (as it is owned by root). 

I gave it the benefit of the doubt that I don't have all the dependencies and did the following:

sudo dnf install qemu
^Installs literally everything it needs. Downside, you have 500M worth of qemu stuff on your system.
sudo ./build.sh
^Yeah not the greatest idea, but hey it worked. 

Not sure how that information can be added to the BUILD. So I'll let it to you.