Make sure PyREBoxLayer uses the correct OS and architecture
The PyREBoxLayer constructor previously hardwired the OS to Windows and the architecture to Intel64.
This change uses the arch value that had already been calculated in the constructor and adds a new API method get_os_kind() to retrieve the OS name as either "Windows" or "Linux" and calls that in the constructor.