Use physical address when validating task list.
In some cases, applying the kernel shift against the physical address of the swapper task does not give the virtual address, so when sanity checking the task structure, we need to map the "next->prev" pointer back to the corresponding physical address to get a valid check.
Also, once that prev pointer is validated, it gives us the correct virtual address of the swapper task and we can then use that when populating the process list.