Use alternative functions when proc_exec_connector is not available
proc_exec_connector() and proc_exit_connector() are only present in the kernel when CONFIG_PROC_EVENTS is enabled.
When those functions are not available, we can use alternatives to track process creation and exit - flush_signal_handlers() and do_exit() appear to be suitable for this. flush_signal_handlers() could be called in circumstances other than process creation, but the code in linux_vmi.cpp already copes with this.