Linux dev (#105)
* Add prettytable and profilehooks to requirements.txt

* Avoid errors due to missing symbols in some linux distros.

Certain distros (for example CentOS 5, Debian 6.0) are missing some of these symbols and this would cause an error in the code.

* Make sure PyREBoxLayer uses the correct OS and architecture

The PyREBoxLayer constructor previously hardwired the OS to Windows and the architecture to Intel64.
This change uses the arch value that had already been calculated in the constructor and adds a new API method get_os_kind() to retrieve the OS name as either "Windows" or "Linux" and calls that in the constructor.

* Hacky fix for linux to force use of first linux banner

Without this change, Linux does not appear to be found by LintelStacker.
However, to use this change, the user must make sure that the only banner/symbols present are the ones relevant to the distro being booted; i.e. when changing guest OS, before starting the new guest, it is necessary to first:
rm -f ${HOME}/.cache/volatility3/linux_banners.cache

* Python3 ascii_lowercase fix for custom modules

The python2 string.lowercase method has been renamed string.ascii_lowercase in python3.

* Revert "Hacky fix for linux to force use of first linux banner"

This reverts commit cca9fb18aa621c7af0ba375a47c706ee30397ca0.

* Avoid infinite loop during search for swapper

In the case when the sanity checks fail after finding the swapper/0 string, offset was not incremented, resulting in an infinite loop

* Only add the process create/delete callbacks when the symbols are present

* Use physical address when validating task list.

In some cases, applying the kernel shift against the physical address of the swapper task does not give the virtual address, so when sanity checking the task structure, we need to map the "next->prev" pointer back to the corresponding physical address to get a valid check.
Also, once that prev pointer is validated, it gives us the correct virtual address of the swapper task and we can then use that when populating the process list.

* Avoid frequent calls to initialize_init_task

Without this change, some OS variants cause initialize_init_task to be called so frequently that little progress is made when booting.
This change adds a further check that initialize_init_task has not been called within n seconds of the previous call, where n starts at 1 and increases by 5 seconds after each call up to a maximum of 120 seconds.

* Use alternative functions when proc_exec_connector is not available

proc_exec_connector() and proc_exit_connector() are only present in the kernel when CONFIG_PROC_EVENTS is enabled.
When those functions are not available, we can use alternatives to track process creation and exit - flush_signal_handlers() and do_exit() appear to be suitable for this. flush_signal_handlers() could be called in circumstances other than process creation, but the code in linux_vmi.cpp already copes with this.

* Tidy up changes to linux_vmi_tlb_callback

Address comments on https://github.com/paulc364/pyrebox/commit/ca9945e9bc500bc652fc40283f5aceb51a26f4e2 by removing unused line and uncommenting another.

* Tidy up addition of process create/delete callbacks

Added some if statements so that we check the create/delete connectors separately and only add callbacks for the alternative functions when they are available.

* Allow Linux swapper task to be found when name is swapper as well as swapper/0

Co-authored-by: xabiugarte <5371486+xabiugarte@users.noreply.github.com>